refactor: factor out relative link click handler, add to panel
This handler makes <a href> relative links use the router instead of incurring a full page reload, which makes them much faster to the user. It is useful to use anywhere that Markdown is rendered.
