+++ src/docSite/DocSitePage.tsx
@@ -14,6 +14,7 @@ import { Markdown } from '../components/Markdown'
 import { PageTitle } from '../components/PageTitle'
 import { eventLogger } from '../tracking/eventLogger'
 import { asError, createAggregateError, ErrorLike, isErrorLike } from '../util/errors'
+import { createLinkClickHandler } from '../util/linkClickHandler'
 import { memoizeObservable } from '../util/memoize'
 
 const queryDocPage = memoizeObservable(
@@ -146,7 +147,7 @@ export class DocSitePage extends React.PureComponent<Props, State> {
                 ) : (
                     <div className="doc-site-page container px-0 d-flex flex-column flex-grow-1">
                         <div
-                            onClick={this.onContentClick}
+                            onClick={createLinkClickHandler(this.props.history)}
                             className="doc-site-page__page flex-grow-1 d-flex flex-wrap flex-row-reverse"
                         >
                             <PageTitle
@@ -178,42 +179,4 @@ export class DocSitePage extends React.PureComponent<Props, State> {
             </>
         )
     }
-
-    private onContentClick: React.MouseEventHandler<HTMLElement> = event => {
-        // Capture clicks on relative links and use pushState for them instead of incurring a full
-        // page reload.
-        if (
-            !event.defaultPrevented &&
-            event.button === 0 &&
-            !event.metaKey &&
-            !event.altKey &&
-            !event.ctrlKey &&
-            !event.shiftKey
-        ) {
-            // Find nearest ancestor <a>.
-            let e: HTMLElement | null = event.target as HTMLElement
-            while (e) {
-                const href = e.getAttribute('href')
-                if (isAnchor(e) && !e.target && href && !/^(https?:)?\/\//.test(href)) {
-                    event.preventDefault()
-                    const url = new URL(e.href)
-                    this.props.history.push({ pathname: url.pathname, hash: url.hash })
-
-                    // HACK: Navigate to the in-page anchor. It does not work without this. This is definitely not
-                    // the best solution. See RenderedFile for another solution (note that using RenderedFile in
-                    // this component's render method instead of Markdown does not work either).
-                    if (url.hash.startsWith('#')) {
-                        setTimeout(() => (window.location.href = url.hash))
-                    }
-
-                    return
-                }
-                e = e.parentElement
-            }
-        }
-    }
-}
-
-function isAnchor(e: HTMLElement): e is HTMLAnchorElement {
-    return e.tagName === 'A'
 }
